+++ lib/Formatting/Interfaces.ml
@@ -1,5 +1,5 @@
-module type CONVERTIBLE = sig
-  (** [CONVERTIBLE] is the interface for types that can be
+module type TO_ELEMENT = sig
+  (** [TO_ELEMENT] is the interface for types that can be
       transformed into a formatting element. *)
 
   type t