+++ CHANGES.md
@@ -4,7 +4,7 @@
 
 - Add `Custom` token type variant which takes a styling, for tokens without particular semantics.
 - Add `Formatting.Element` that supersedes `Formatting.Tree`.
-- Add `Formatting.Interfaces.CONVERTIBLE` interface which establishes the contract to convert to a formatting element that is used by formatting and printing utilitary functions such as `format` and `IO.print_formatted`.
+- Add `Formatting.Interfaces.TO_ELEMENT` interface which establishes the contract to convert to a formatting element that is used by formatting and printing utilitary functions such as `format` and `IO.print_formatted`.
 - Expose the `Int8` module that is used by `Color`.
 
 ## Removed